refactor: extract shared binary resolver
