+++ src/binary_resolver/inspect.js
